Update README to cover multi-nodes in kitchen.yml
This patch updates the README to clarify the ability of a user to create and provision multiple nodes via test-kitchen and this driver. It shows the users how to overwrite driver settings in both the platform and suite sections. Finally, it gives the helpful tip for users to call `kitchen diagnose` to verify their settings prior to `kitchen create` execution.